requirements.txt
Flask==0.11.1
azure-storage==0.32.0
pyodbc==3.0.10
